Moved content of custom_types.filters.py to filters module
