+++ backend/subprocess_main.py
@@ -5,7 +5,7 @@ import sys
 from typing import Tuple
 from aiortc import RTCSessionDescription
 
-from custom_types.filters import FilterDict
+from filters import FilterDict
 from custom_types.connection import is_valid_rtc_session_description_dict
 
 from modules.connection_runner import ConnectionRunner
